--- a/Gemfile
+++ b/Gemfile
@@ -99,3 +99,11 @@ group :development, :test do
   gem "binding_of_caller"
   gem 'ruby-prof'
 end
+
+group :development do
+  # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
+  gem 'spring'
+  gem 'spring-commands-rspec'
+  gem 'spring-watcher-listen', '~> 2.0.0'
+  gem 'listen', '~> 3.0.5'
+end