--- a/app/controllers/recent_controller.rb
+++ b/app/controllers/recent_controller.rb
@@ -14,7 +14,7 @@ class RecentController < ApplicationController
   end
 
   protected
-  
+
     def recent
       params.permit(:year, :month)
       # grab any recent documents
@@ -22,7 +22,7 @@ class RecentController < ApplicationController
     end
 
     def date_buckets
-      solr_rsp = ActiveFedora::SolrService.instance.conn.get "select", :params => {:q => "#{Solrizer.solr_name('read_access_group', :symbol)}:public #{Solrizer.solr_name('active_fedora_model', :stored_sortable)}:GenericFile", :fq => "-#{Solrizer.solr_name('read_access_group', :symbol)}:#{Hydranorth::AccessControls::InstitutionalVisibility::UNIVERSITY_OF_ALBERTA}", 'facet.range' => "#{Solrizer.solr_name('system_create', :stored_sortable, type: :date)}", 'facet.range.gap' => '+1MONTH', 'facet.range.start' => '1906-01-01T00:00:00Z', 'facet.range.end' => 'NOW', :rows => 0, 'facet.mincount' => 1 } 
+      solr_rsp = ActiveFedora::SolrService.instance.conn.get "select", :params => {:q => "#{Solrizer.solr_name('read_access_group', :symbol)}:public #{Solrizer.solr_name('active_fedora_model', :stored_sortable)}:GenericFile", :fq => "-#{Solrizer.solr_name('read_access_group', :symbol)}:#{Hydranorth::AccessControls::InstitutionalVisibility::UNIVERSITY_OF_ALBERTA}", 'facet.range' => "#{Solrizer.solr_name('system_create', :stored_sortable, type: :date)}", 'facet.range.gap' => '+1MONTH', 'facet.range.start' => '1906-01-01T00:00:00Z', 'facet.range.end' => 'NOW', :rows => 0, 'facet.mincount' => 1 }
       range_counts = solr_rsp['facet_counts']['facet_ranges']['system_create_dtsi']['counts']
       @date_buckets = {}
       range_counts.each_slice(2) { |date,count| (@date_buckets[date.slice(0,4)] ||= []) << [date,count] }
@@ -49,6 +49,7 @@ class RecentController < ApplicationController
       flash[:alert] = "Couldn't interpret the date."
     end
 
+    # TODO: show_only_public_files is deprecated
     def show_only_public_files(solr_parameters, user_parameters)
       solr_parameters[:fq] ||= []
       solr_parameters[:fq] += ["-#{Solrizer.solr_name('read_access_group', :symbol)}:#{Hydranorth::AccessControls::InstitutionalVisibility::UNIVERSITY_OF_ALBERTA}"]