--- a/config/application.rb
+++ b/config/application.rb
@@ -30,6 +30,6 @@ module Hydranorth
     # config.i18n.default_locale = :de
 
     # Do not swallow errors in after_commit/after_rollback callbacks.
-    #config.active_record.raise_in_transactional_callbacks = true
+    config.active_record.raise_in_transactional_callbacks = true
   end
 end