… Gemfile.lock
… lib/tasks/migration.rake
… lib/tasks/migration_audit.rake
… spec/controllers/communities_controller_spec.rb
… spec/controllers/my/files_controller_spec.rb
… spec/controllers/recent_controller_spec.rb
… spec/controllers/redirect_controller_spec.rb
… spec/controllers/selects_collections_controller_spec.rb
… spec/controllers/single_use_links_viewer_controller_spec.rb
… spec/controllers/users/omniauth_callbacks_controller.rb
… spec/controllers/users_controller_spec.rb
… spec/controllers/users_spec.rb
… spec/factories/collections_factory.rb
… spec/factories/generic_files.rb
… spec/factories/users.rb
… spec/features/admin_delete_spec.rb
… spec/features/admin_tasks_spec.rb
… spec/features/advanced_search_spec.rb
… spec/features/agreement_spec.rb
… spec/features/analytics_spec.rb
… spec/features/catalog_search_spec.rb
… spec/features/coauthor_spec.rb
… spec/features/collection/1008_users_shouldnt_see_edit_collection_spec.rb
… spec/features/collection/collection_spec.rb
… spec/features/community_spec.rb
… spec/features/display_dashboard_spec.rb
… spec/features/display_home_page_spec.rb
… spec/features/doi_spec.rb
… spec/features/edit_file_spec.rb
… spec/features/edit_license_spec.rb