This pull request contains a number of refactors for our test suite. The main goal here is for speed improvements. Previously a full test suite run using master was taking over > 25-30 minutes on my local machine. Here are some timings I got from today on master branch on 427 examples:
- 33 minutes and 46 seconds
- 36 minutes and 46 seconds
- 31 minutes and 29 seconds

After this pull request, it is around ~16 minutes on my local. Here are some timing I have gotten on this PR on 335 examples:
- 16 minutes and 24 seconds
- 15 minutes and 48 seconds
- 15 minutes and 26 seconds
- 15 minutes and 37 seconds
- 16 minutes and 46 seconds

After merging this with master (which has my code cleanup commits) I am getting around ~11 mins, Here are some timings I have gotten:
- 11 minutes 33 seconds
- 11 minutes 29 seconds
- 11 minutes 30 seconds
- 11 minutes 33 seconds

Root cause? Seems to be `better_errors` being in the test group and running in our tests! Nice 4 minute speedup! Interesting.

I assume this will be a vast speed up on our VM's as well.

I only really looked deeply into the controller tests and refactored those, so there is definitely more improvements that could be made especially for the integration/feature test which consumes over half the total speed time! (Most of the folders (controllers/models/jobs/etc) take around 1-2 mins each, but `/features` consumes almost 9-10 minutes of the total time)

Ultimately the biggest bottleneck is fedora. So doing this smarter will net the biggest gains. Simply using a `GenericFile.new` instead of a `GenericFile.create` can net an easy 10 seconds! So rule of thumb is don't create an object when `build`, `build_stubbed`, `attributes_for`, `spy`, or `double` will do. Database persistence is slow (especially in fedora)! We might want to start mocking more?

On the fedora note, I think we should rethink on how we are using FactoryGirl, if at all for fedora. FactoryGirl only works well when you are using transactions. For a normal database, FactoryGirl will create the object in a transaction, run the example, then rollback the transaction. So it will always cleanup after itself if everything is in a transaction. If you use FactoryGirl in an `before(:all)` you will notice it will dirty the database as it's no longer in the transaction for an example. Fedora is the same way, it does not support transactions. So it will always create objects but never clean them up. So if you have something like the following:

```ruby
let!(:file) do
  FactoryGirl.create(:generic_file)
end
```
 And you had 9 examples in your spec. It will leave 9 objects behind. This is not ideal at all and not the expected behaviour when using FactoryGirl! This is big reason we have `cleanup_jetty` or `GenericFile.delete_all`, throughout the test suite in the before and after blocks which is a pretty big code smell. 

In regards to cleaning up your data, always use the `cleanup_jetty` method which is just calling `ActiveFedora::Cleaner.clean!`. Running some benchmarks on a typical spec you will see cleanup_jetty is much much faster then other cleanup methods. Which seems weird? Maybe someone with some knowledge of fedora can explain this?

Say you are creating 3 Generic files in a before each block:

```ruby
before do
  @gf1 = FactoryGirl.create(:genericfile)
  @gf2 = FactoryGirl.create(:genericfile)
  @gf3 = FactoryGirl.create(:genericfile)
end
```
And then you might clean up these objects in an after each block,
```ruby
after do
  # Could just called GenericFile.delete_all or GenericFile.destroy_all
  # Could just call cleanup_jetty
  @gf1.delete
  @gf2.delete
  @gf3.delete
end
```
On a typical spec using `.delete` on each object  I was getting ~6 seconds for the above (6.13, 6.25 and 5.96). Using `GenericFile.delete_all` I was getting ~8 seconds (7.82, 7.7, 8.29). And finally using `cleanup_jetty` I was getting ~5 seconds (4.55, 4.53, 5.23). So lesson is, use `cleanup_jetty` where possible.

The other big issues outside of fedora is just smarter testing in general. We are testing the wrong things (tests to make sure underlining gems, ruby/rails, or rspec are working correctly shouldn't be needed),  have useless test setup's, lot's of copy pasta, etc. I hardly touched the big bottleneck (fedora) and simply just refactoring our tests to be smarter shaved off alot of minutes on our total test time! It's important that we make an effort to write tests that are accurate and effective as well as fast.


Other things to look out for in the future:

- data stores (such as database, job queues, ActionMailer::Base.deliveries, etc) should return to a consistent state between each spec, regardless of the order specs are run in.
  - cleaning database before and after every example shouldn't be needed? (via database_cleaner or cleanup_jetty)
  - tests should be configured to run with seeds/random order should help with this
- Use cleanup_jetty vs delete/delete_all/destroy_all
- Don't mark a feature as requiring JavaScript (through js: true in RSpec) unless it's actually required for the test to be valid. Headless browser testing is slow!
- Should be using `rails_helper.rb` for rails code  and `spec_helper.rb` (light weight version) for non rails code

Code style things:
- don't use `.tap` on factorygirl/activerecord/activefedora objects (let's not use blocks in general? If you are using blocks, `.new/.build` then call save in the block! Do not use `.create` and then call `.save` inside the block! We seem to abusing them because we use `#apply_depositor_metadata` helper method all over the place. Which from my understand just sets the `depositor_id` and `edit_users` with user's email?  http://www.rubydoc.info/gems/hydra-access-controls/10.0.1/Hydra%2FWithDepositor%3Aapply_depositor_metadata)
- Try not to use `subject` in rspec, especially implicit
- Don't use one liner `it {}` syntax if not in isolation
- Pick either `before` or `before(:each)` syntax and use everywhere. `:each` is the default. I would prefer to be specific about it, as I've seen a few cases where we think it is a `before(:all)`. Regardless be consistent about it.

Matt makes some great recommendations in the original issue as well: https://github.com/ualbertalib/HydraNorth/issues/778

Some resources from around the web:
- http://betterspecs.org/
- Thoughtbot's blog has a ton of great testing articles: 
  - https://robots.thoughtbot.com/speed-up-tests-by-selectively-avoiding-factory-girl
  - https://robots.thoughtbot.com/four-phase-test
  - etc.
- https://github.com/thoughtbot/guides/tree/master/style/testing
- https://gitlab.com/gitlab-org/gitlab-ce/blob/master/doc/development/testing.md
- https://github.com/eliotsykes/rspec-rails-examples