@pgwillia 2016-12-06
This is awesome! (So speedy) But why are we seeing a different number of examples? And is this only failing for me?

Failures:

  1) SelectsCollectionsController Select Communities Regular User Read Access not signed in should error if the user is not signed in
     Failure/Error: expect { subject.find_communities([:read]) }.to raise_error(UncaughtThrowError, 'uncaught throw :warden')
     NameError:
       uninitialized constant UncaughtThrowError
     # ./spec/controllers/selects_collections_controller_spec.rb:58:in `block (5 levels) in <top (required)>'

  2) SelectsCollectionsController Select Communities Regular User Edit Access not signed in should error if the user is not signed in
     Failure/Error: expect { subject.find_communities([:edit])  }.to raise_error(UncaughtThrowError, 'uncaught throw :warden')
     NameError:
       uninitialized constant UncaughtThrowError
     # ./spec/controllers/selects_collections_controller_spec.rb:73:in `block (5 levels) in <top (required)>'

Finished in 11 minutes 52 seconds (files took 11.15 seconds to load)
335 examples, 2 failures, 6 pending