@mbarnett 2016-12-06
Oh, I had the migration spec set up to skip, on further investigation

What version of ruby are you using? UncaughtThrowError is new in 2.2, I think. 

We should probably update our gemfile to reflect a minimum ruby version.