@pgwillia 2016-12-06
That's likely it.  I see that ruby 2.1.5 is what was using when running these specs.