@murny 2016-12-06
The different amount of examples is due to refactoring. Lots of places where we were testing the exact same thing but having it as different examples (which means before/after blocks run more times then needed). For example models/ability_spec.rb (https://github.com/ualbertalib/HydraNorth/pull/1316/files#diff-e9501dda6c23fd3e1076e2dff66d0f68), has 84 examples and is now down to under 10 as I have refactor this to get away from the one liner `it{}` syntax which should never be used when you are not in isolation.