@mbarnett 2016-12-06
Any objections to adding a Ruby 2.2 minimum version declaration?